few fixed and simple install script

+++ launcher.sh
@@ -1,14 +1,15 @@
 #!/bin/bash -
 
-echo "Kill existing pid and remove daemon pid -files in /tmp/"
-
 pid=$(pgrep -f "no-ip-daemon.py")
 
 if [[ $pid = *[[:digit:]]* ]]; then
     echo "Found process with pid ("$pid"). Killing it."
     kill -9 $pid
 fi
-rm /tmp/no-ip-daemon.pid
+if [ -f "/tmp/no-ip-daemon.pid" ]; then
+    echo "Found pid file. Removing."
+    rm /tmp/no-ip-daemon.pid
+fi
 
 cd /opt/no-ip-daemon
 python no-ip-daemon.py
